Added OpenMP directives to the FPT.
Also added a performance utest to check the implementation.
Note: this first implementation with OpenMP breaks for polynomial
degrees of 255 and higher (the algorithm starts to use approximated
blocks then).
This is probably due to the usage of memory allocated in the
initialization for the FPT, that is shared between threads.
There is no problem when using OMP_NUM_THREADS=1 for the
ply_fpt_3D_performance_test.

--HG--
branch : feature/openmp