Added a decrement dof reducing method to the adaptive subsampling.
This mode allows us to always cut off a fixed number of highest modes
on each refinement level.
The change introduces a new setting 'reduction_mode', which can either
be 'factor' or 'decrement', which defaults to 'factor' to be compatible
with previous configurations.
If the mode is set to decrement, the 'dof_decrement' option can be used
to specify how many modes should be cut off in each refinement.
It defaults to 1.